Final polish of ActivityGraphView rendering and GPUView placeholder.
- Fixed visual discontinuity in ActivityGraphView partial updates by using fLastRefresh for all data points.
- Improved precision and robustness of grid line drawing in ActivityGraphView by using float for fScrollOffset.
- Refactored GPUView utilization display to use a member BStringView and added explanatory comments.
- Cleaned up unused variables and redundant logic in ProcessView and Utils.

+++ SystemMonitor/ActivityGraphView.h
@@ -40,7 +40,7 @@ private:
 	int64				fManualMax;
 
 	bigtime_t			fLastRefresh;
-	int64				fScrollOffset;
+	float				fScrollOffset;
 };
 
 #endif // ACTIVITYGRAPHVIEW_H
